refactor: share browser session implementation
