+++ src/browser/local-provider.ts
@@ -1,68 +1,23 @@
-import { Buffer } from 'node:buffer';
-import { assertBrowserNavigationUrl } from '../../container/shared/browser-navigation.js';
 import { BROWSER_PROFILE_CHROMIUM_ARGS } from '../../container/shared/browser-profile.js';
 import type { SecretHandle } from '../security/secret-handles.js';
-import type { SecretInput } from '../security/secret-refs.js';
 import {
-  fillBrowserField,
   loadPlaywrightModule,
-  normalizeScrollDelta,
-  type PlaywrightNavigationOptions,
-  type PlaywrightSecretFillLocator,
-  toNavigationOptions,
+  PlaywrightBrowserSession,
+  type PlaywrightContextShape,
+  type PlaywrightPageShape,
 } from './playwright-utils.js';
 import {
   resolveBrowserProfileRoot,
   resolveConstrainedBrowserProfileDir,
 } from './profile-dir.js';
 import type {
-  BrowserEvaluateFunction,
   BrowserProvider,
   BrowserSession,
-  ClickOptions,
-  HistoryNavigationOptions,
-  NavigateOptions,
-  ScreenshotOptions,
-  ScrollOptions,
   SessionOptions,
-  WaitOptions,
 } from './provider.js';
 
-type PlaywrightScreenshotOptions = {
-  fullPage?: boolean;
-  type?: 'png' | 'jpeg';
-};
-
-type PlaywrightPage = {
-  evaluate<T>(fn: BrowserEvaluateFunction<T>): Promise<T>;
-  screenshot(opts?: PlaywrightScreenshotOptions): Promise<Buffer | Uint8Array>;
-  goto(url: string, opts?: PlaywrightNavigationOptions): Promise<unknown>;
-  goBack(opts?: PlaywrightNavigationOptions): Promise<unknown>;
-  goForward(opts?: PlaywrightNavigationOptions): Promise<unknown>;
-  reload(opts?: PlaywrightNavigationOptions): Promise<unknown>;
-  click(selector: string, opts?: { timeout?: number }): Promise<void>;
-  fill(selector: string, value: string): Promise<void>;
-  url(): string;
-  mouse: {
-    wheel(deltaX: number, deltaY: number): Promise<void>;
-  };
-  waitForSelector(
-    selector: string,
-    opts?: { state?: WaitOptions['state']; timeout?: number },
-  ): Promise<unknown>;
-  locator(selector: string): PlaywrightSecretFillLocator & {
-    evaluate<TArg>(
-      fn: (element: Element, arg: TArg) => void,
-      arg: TArg,
-    ): Promise<void>;
-  };
-};
-
-type PlaywrightContext = {
-  pages(): PlaywrightPage[];
-  newPage(): Promise<PlaywrightPage>;
-  close(): Promise<void>;
-};
+type PlaywrightPage = PlaywrightPageShape;
+type PlaywrightContext = PlaywrightContextShape<PlaywrightPage>;
 
 export type LocalBrowserPlaywrightModule = {
   chromium: {
@@ -78,6 +33,7 @@ export type LocalBrowserPlaywrightModule = {
 };
 
 export interface LocalBrowserProviderOptions {
+  /** Test/direct-construction fallback; production config passes profileRoot. */
   dataDir?: string;
   profileRoot?: string;
   headed?: boolean;
@@ -95,80 +51,7 @@ async function loadPlaywright(
   );
 }
 
-class LocalBrowserSession implements BrowserSession {
-  constructor(
-    private readonly page: PlaywrightPage,
-    private readonly secretAudit?: (
-      handle: SecretHandle,
-      reason: string,
-    ) => void,
-    private readonly metering?: SessionOptions['metering'],
-  ) {}
-
-  async evaluate<T>(fn: BrowserEvaluateFunction<T>): Promise<T> {
-    return await this.page.evaluate(fn);
-  }
-
-  async screenshot(opts?: ScreenshotOptions): Promise<Buffer> {
-    const bytes = await this.page.screenshot({
-      fullPage: opts?.fullPage,
-      type: opts?.type,
-    });
-    return Buffer.isBuffer(bytes) ? bytes : Buffer.from(bytes);
-  }
-
-  async navigate(url: string, opts?: NavigateOptions): Promise<void> {
-    const parsed = await assertBrowserNavigationUrl(url);
-    await this.page.goto(parsed.toString(), toNavigationOptions(opts));
-  }
-
-  async back(opts?: HistoryNavigationOptions): Promise<void> {
-    await this.page.goBack(toNavigationOptions(opts));
-  }
-
-  async forward(opts?: HistoryNavigationOptions): Promise<void> {
-    await this.page.goForward(toNavigationOptions(opts));
-  }
-
-  async reload(opts?: HistoryNavigationOptions): Promise<void> {
-    await this.page.reload(toNavigationOptions(opts));
-  }
-
-  async click(selector: string, opts?: ClickOptions): Promise<void> {
-    await this.page.click(selector, { timeout: opts?.timeoutMs });
-  }
-
-  async fill(selector: string, value: SecretInput): Promise<void> {
-    await fillBrowserField(
-      this.page,
-      selector,
-      value,
-      this.secretAudit,
-      this.metering,
-    );
-  }
-
-  async scroll(opts: ScrollOptions): Promise<void> {
-    const delta = normalizeScrollDelta(opts);
-    if (opts.selector) {
-      await this.page
-        .locator(opts.selector)
-        .evaluate((element, scrollDelta) => {
-          element.scrollBy(scrollDelta.deltaX, scrollDelta.deltaY);
-        }, delta);
-      return;
-    }
-
-    await this.page.mouse.wheel(delta.deltaX, delta.deltaY);
-  }
-
-  async waitForSelector(selector: string, opts?: WaitOptions): Promise<void> {
-    await this.page.waitForSelector(selector, {
-      state: opts?.state,
-      timeout: opts?.timeoutMs,
-    });
-  }
-}
+class LocalBrowserSession extends PlaywrightBrowserSession<PlaywrightPage> {}
 
 export class LocalBrowserProvider implements BrowserProvider {
   private readonly profileRoot: string;
