+++ src/browser/playwright-utils.ts
@@ -1,3 +1,5 @@
+import { Buffer } from 'node:buffer';
+import { assertBrowserNavigationUrl } from '../../container/shared/browser-navigation.js';
 import { makeAuditRunId, recordAuditEvent } from '../audit/audit-events.js';
 import {
   assertSecretResolveAllowed,
@@ -14,15 +16,56 @@ import {
   type SecretRef,
 } from '../security/secret-refs.js';
 import type {
+  BrowserEvaluateFunction,
+  BrowserSession,
   BrowserSessionMeteringContext,
+  ClickOptions,
   HistoryNavigationOptions,
   NavigateOptions,
+  ScreenshotOptions,
   ScrollOptions,
+  WaitOptions,
 } from './provider.js';
+
+export type PlaywrightScreenshotOptions = {
+  fullPage?: boolean;
+  type?: 'png' | 'jpeg';
+};
 
 export type PlaywrightNavigationOptions = {
   waitUntil?: 'load' | 'domcontentloaded';
   timeout?: number;
+};
+
+export type PlaywrightPageShape = {
+  evaluate<T>(fn: BrowserEvaluateFunction<T>): Promise<T>;
+  screenshot(opts?: PlaywrightScreenshotOptions): Promise<Buffer | Uint8Array>;
+  goto(url: string, opts?: PlaywrightNavigationOptions): Promise<unknown>;
+  goBack(opts?: PlaywrightNavigationOptions): Promise<unknown>;
+  goForward(opts?: PlaywrightNavigationOptions): Promise<unknown>;
+  reload(opts?: PlaywrightNavigationOptions): Promise<unknown>;
+  click(selector: string, opts?: { timeout?: number }): Promise<void>;
+  fill(selector: string, value: string): Promise<void>;
+  url(): string;
+  mouse: {
+    wheel(deltaX: number, deltaY: number): Promise<void>;
+  };
+  waitForSelector(
+    selector: string,
+    opts?: { state?: WaitOptions['state']; timeout?: number },
+  ): Promise<unknown>;
+  locator(selector: string): PlaywrightSecretFillLocator & {
+    evaluate<TArg>(
+      fn: (element: Element, arg: TArg) => void,
+      arg: TArg,
+    ): Promise<void>;
+  };
+};
+
+export type PlaywrightContextShape<TPage extends PlaywrightPageShape> = {
+  pages(): TPage[];
+  newPage(): Promise<TPage>;
+  close(): Promise<void>;
 };
 
 export type PlaywrightFillPage = {
@@ -39,6 +82,84 @@ export type PlaywrightSecretFillLocator = {
   fill(value: string): Promise<void>;
   pressSequentially?(value: string): Promise<void>;
 };
+
+export class PlaywrightBrowserSession<
+  TPage extends PlaywrightPageShape = PlaywrightPageShape,
+> implements BrowserSession
+{
+  constructor(
+    private readonly page: TPage,
+    private readonly secretAudit?: (
+      handle: SecretHandle,
+      reason: string,
+    ) => void,
+    private readonly metering?: BrowserSessionMeteringContext,
+  ) {}
+
+  async evaluate<T>(fn: BrowserEvaluateFunction<T>): Promise<T> {
+    return await this.page.evaluate(fn);
+  }
+
+  async screenshot(opts?: ScreenshotOptions): Promise<Buffer> {
+    const bytes = await this.page.screenshot({
+      fullPage: opts?.fullPage,
+      type: opts?.type,
+    });
+    return Buffer.isBuffer(bytes) ? bytes : Buffer.from(bytes);
+  }
+
+  async navigate(url: string, opts?: NavigateOptions): Promise<void> {
+    const parsed = await assertBrowserNavigationUrl(url);
+    await this.page.goto(parsed.toString(), toNavigationOptions(opts));
+  }
+
+  async back(opts?: HistoryNavigationOptions): Promise<void> {
+    await this.page.goBack(toNavigationOptions(opts));
+  }
+
+  async forward(opts?: HistoryNavigationOptions): Promise<void> {
+    await this.page.goForward(toNavigationOptions(opts));
+  }
+
+  async reload(opts?: HistoryNavigationOptions): Promise<void> {
+    await this.page.reload(toNavigationOptions(opts));
+  }
+
+  async click(selector: string, opts?: ClickOptions): Promise<void> {
+    await this.page.click(selector, { timeout: opts?.timeoutMs });
+  }
+
+  async fill(selector: string, value: SecretInput): Promise<void> {
+    await fillBrowserField(
+      this.page,
+      selector,
+      value,
+      this.secretAudit,
+      this.metering,
+    );
+  }
+
+  async scroll(opts: ScrollOptions): Promise<void> {
+    const delta = normalizeScrollDelta(opts);
+    if (opts.selector) {
+      await this.page
+        .locator(opts.selector)
+        .evaluate((element, scrollDelta) => {
+          element.scrollBy(scrollDelta.deltaX, scrollDelta.deltaY);
+        }, delta);
+      return;
+    }
+
+    await this.page.mouse.wheel(delta.deltaX, delta.deltaY);
+  }
+
+  async waitForSelector(selector: string, opts?: WaitOptions): Promise<void> {
+    await this.page.waitForSelector(selector, {
+      state: opts?.state,
+      timeout: opts?.timeoutMs,
+    });
+  }
+}
 
 type CredentialFillParams = {
   selector: string;
